mcp/github_server/requirements.txt
fastmcp==2.14.4
httpx>=0.28.1
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
PyYAML>=6.0
PyGithub>=2.5.0